pyproject.toml
[build-system]
requires = ["flit-core~=3.9"]
build-backend = "flit_core.buildapi"

[project]
name = "flood-client"
version = "0.0.1"
description = "flood.js python client"
authors = [
    { name = "trim21", email = "trim21me@gmail.com" },
]
readme = 'readme.md'
license = { text = 'MIT' }
keywords = ['flood', 'rpc']
classifiers = [
    'Intended Audience :: Developers',
    'Development Status :: 4 - Beta',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python :: 3 :: Only',
]

requires-python = "~=3.8"

dependencies = [
    'sseclient-py>=1.8.0,<2',
    'httpx>=0.27.0,<1',
    'typing-extensions>=4.7.1',
]

[project.optional-dependencies]
orjson = [
    'orjson~=3.0'
]

docs = [
    'sphinx~=8.0; python_version >= "3.9"',
    'furo~=2024.0; python_version >= "3.9"',
]

dev = [
    "pytest==8.3.2",
    "pytest-github-actions-annotate-failures==0.2.0",
    "coverage==7.6.1",
    'pre-commit==3.8.0; python_version >= "3.9"',
    'sphinx-autobuild==2024.4.16; python_version >= "3.9"',
    'mypy==1.11.1; python_version >= "3.9"',
]

[project.urls]
Homepage = "https://github.com/trim21/flood-client-py"

[tool.pytest.ini_options]
addopts = '-rav -Werror'

[tool.mypy]
python_version = "3.8"
disallow_untyped_defs = true
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true
show_error_codes = true
platform = 'unix'

[tool.black]
target-version = ['py38']

[tool.ruff]
target-version = "py38"

[tool.ruff.lint]
select = [
    "B",
    "C",
    "E",
    "F",
    "G",
    "I",
    "N",
    "Q",
    "S",
    "W",
    "BLE",
    "EXE",
    "ICN",
    "INP",
    "ISC",
    "NPY",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PYI",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "TCH",
    "TID",
    "TRY",
    "YTT",
]

ignore = [
    'PLR0911',
    'INP001',
    'N806',
    'N802',
    'N803',
    'E501',
    'BLE001',
    'RUF002',
    'S324',
    'S301',
    'S314',
    'S101',
    'N815',
    'S104',
    'C901',
    'PLR0913',
    'RUF001',
    'SIM108',
    'TCH003',
    'RUF003',
    'RET504',
    'TRY300',
    'TRY003',
    'TRY201',
    'TRY301',
    'PLR0912',
    'PLR0915',
    'PLR2004',
    'PGH003',
]
